+++ common/Dto/vehicle_wall_collision.h
@@ -8,8 +8,8 @@
 class VehicleWallCollisionDto: public Dto {
 public:
     std::string username;
-    int vehicle_new_hp;
-    VehicleWallCollisionDto(const std::string& username, int vehicle_new_hp);
+    float vehicle_new_hp;
+    VehicleWallCollisionDto(const std::string& username, float vehicle_new_hp);
 };
 
 #endif
